pyproject.toml
[project]
name = "remnastore"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.6.0",
    "asyncpg>=0.29.0",
    "fastapi>=0.111.0",
    "httpx>=0.27.0",
    "pydantic-settings>=2.3.0",
    "redis>=5.2.1",
    "remnawave>=2.6.3",
    "sqlalchemy[asyncio]>=2.0.0",
    "uvicorn[standard]>=0.30.0",
    "alembic>=1.13.3",
    "yookassa==3.10.0",
    "aiosqlite>=0.22.1",
]

[dependency-groups]
dev = [
    "coverage[toml]>=7.10.0,<8",
    "ruff>=0.15.7",
]

[tool.coverage.run]
branch = true
relative_files = true

[tool.coverage.report]
show_missing = true
skip_covered = false
skip_empty = true
omit = [
    "*/__init__.py",
]
